Add instrs for deploying on external machine
Two lines in the web-app must be changed before deploying the Swagger
Web-UI. If these lines are not changed, the user will see the
following message, as the server would attempt to search the localhost
instead of the Swagger machine:

can't read from server. It may not have the appropriate
access-control-origin settings.